refactor: 출석 상태 통계 갱신 책임을 StatusStatistic로 이동
