+++ src/test/java/attendance/domain/AttendancesTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 import java.time.LocalDateTime;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -45,24 +44,5 @@ class AttendancesTest {
         assertThatThrownBy(() -> attendances.addAttendance(duplicate))
             .isInstanceOf(AttendanceArgumentException.class)
             .hasMessageContaining("이미 출석하였습니다. 수정 기능을 이용해주세요.");
-    }
-
-    @Test
-    @DisplayName("출석 상태 통계를 반환한다.")
-    void test_returnAttendanceStatusStatistic() {
-        var attendance = LocalDateTime.of(2024, 12, 10, 10, 0);
-        var late = LocalDateTime.of(2024, 12, 11, 10, 10);
-        var absence = LocalDateTime.of(2024, 12, 12, 10, 35);
-        var attendances = new Attendances(systemDateTime);
-        attendances.addAttendance(attendance);
-        attendances.addAttendance(late);
-        attendances.addAttendance(absence);
-
-        StatusStatistics statusStatics = attendances.calculateStatics();
-        Assertions.assertAll(
-            () -> assertThat(statusStatics.get(AttendanceStatus.ATTENDANCE)).isEqualTo(1),
-            () -> assertThat(statusStatics.get(AttendanceStatus.LATE)).isEqualTo(1),
-            () -> assertThat(statusStatics.get(AttendanceStatus.ABSENCE)).isEqualTo(1)
-        );
     }
 }
